src/requirements.txt
Django~=5.0.3
Django-allauth
Django-role-permissions
Django-template-partials
python-dotenv==1.0.1
Django-debug-toolbar
Pillow~=10.3.0